+++ CopilotKit/packages/runtime/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @copilotkit/runtime
 
+## 1.10.2
+
+### Patch Changes
+
+- @copilotkit/shared@1.10.2
+
 ## 1.10.2-next.0
 
 ### Patch Changes